Fixed comment syntax
