+++ Auth/Auth.Platform.Android/Authenticator/TokenLifecycleManagerExtensions.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using AeroGear.Mobile.Auth.Authenticator;
 using OpenId.AppAuth;
 using Org.Json;
-using AuthException = AeroGear.Mobile.Auth.Authenticator.AuthorizationException;
 
 namespace Auth.Platform.Authenticator.extensions
 {
@@ -15,14 +14,14 @@ namespace Auth.Platform.Authenticator.extensions
                 var json = await tlcm.RefreshAsync(tokenRequest.Configuration.TokenEndpoint.ToString(), tokenRequest.ClientId, tokenRequest.RefreshToken).ConfigureAwait(false);
                 return new TokenResponse.Builder(tokenRequest).FromResponseJson(new JSONObject(json.ToString())).Build();
             }
-            catch (AuthException ae)
+            catch (AuthzException ae)
             {
                 throw ae;
             }
             catch (Exception je)
             {
-                throw AuthException.fromTemplate(
-                    AuthException.GeneralErrors.JSON_DESERIALIZATION_ERROR, je);
+                throw AuthzException.fromTemplate(
+                    AuthzException.GeneralErrors.JSON_DESERIALIZATION_ERROR, je);
             }
         }
     }
